requirements.txt
pip
setuptools
wheel
tensorflow
numpy
pandas
Pillow
scipy
scikit-learn
tensorflow-hub
tensorflow-text
tf-models-official
matplotlib
seaborn
openpyxl
flask
requests
six
librosa
pytesseract
python-docx
pdf2docx
tabula-py
